requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.10.1
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
channels==4.0.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
distlib==0.3.7
Django==4.2.6
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-widget-tweaks==1.5.0
filelock==3.13.1
frozenlist==1.4.0
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
identify==2.5.31
idna==3.4
iniconfig==2.0.0
model-bakery==1.17.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
postgres==4.0
pre-commit==3.5.0
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
yarl==1.9.2
